Esdots now finds random access points in H.264 streams. These are defined as : IDR pic OR I pic with SEI unit with recovery_point. In choose_nal_type in esdotc.c there is a boolean value rec_point_required. It you set it to FALSE any I pic is considered a random access point. The field recovery_frame_cnt is not used (as if it was 0). If it is != 0 a warning message is printed. The fields broken_link_flag, exact_match_flag and changing_slice_group_idc are not considered (as 0).
In the final message the max, min and mean duration of a GOP are shown. For H.264 streams this is defined as the distance between 2 random access points. If no full GOP is found this info is not displayed.

The -gop switch works as before.

In esdots a function choose_nal_type has been added.
In nalunit.c a function read_SEI_recovery_point has been added. 
The structure nal_SEI_recovery_data has been added, in order to cope with the recovery_point message. It's added to nal_innards too. See nalunit_defns.h




git-svn-id: svn://svn.berlios.de/tstools/trunk@9 eff31bef-be4a-0410-a8fe-e47997df2690