Rename es.pyx as tstools.pyx -- mainly because pyrex assumes that the final Python package will have the same name as the pyx file (or nasty things happen when you try to import it), and I ultimately want the Python module to be called "tstools".
Add a (probably) temporary Makefile for pyrex, and the Setup.py that does
the actual work.

On my Mac, this how builds something I can import into Python, and then
discover errors in... so fix the 'filename' attribute of ESStream, both
to exist (!) and secondly to be visible to Python.


git-svn-id: svn://svn.berlios.de/tstools/trunk@25 eff31bef-be4a-0410-a8fe-e47997df2690